+++ Assets/00-5. YeoJin/01.Scenes/Main_copied_test.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 730c2e2c1f7999a44a7ef747cd83ce2d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
