--- Assets/00-5. YeoJin/01.Scenes/TestScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2424c1c24826f8645b3563cfc4ad8531
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
