+++ Assets/00-5. YeoJin/02.Scripts/Player/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9768a17a4576cc244becdff66d31ff2e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
