+++ Assets/00-5. YeoJin/02.Scripts/Player/UI/UI_HealthBar.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+public class UI_HealthBar : MonoBehaviour
+{
+    void Start()
+    {
+        
+    }
+
+    private void Update()
+    {
+        transform.LookAt(transform.position + Camera.main.transform.rotation * Vector3.forward, Camera.main.transform.rotation * Vector3.up);
+    }
+}
