+++ Assets/00-5. YeoJin/02.Scripts/Player/UI/UI_NicknameIngame.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cc0734a0385c1f049b6453d80b1342b3
